use stimulus to work timer
