+++ app/assets/stylesheets/application.scss
@@ -42,7 +42,7 @@ textarea {
 }
 
 
-#time {
+[data-timer-target="time"] {
   text-align: center;
   font-size: 2.5rem;
   margin: .5rem;
